PhoneId retrieved as long
- PhoneIdRetriever class is added in util
- phoneId(memberId) is long type

+++ WhenIsBetter/src/team/gangnamstyle/whenisbetter/model/Project.java
@@ -6,7 +6,7 @@ import java.util.List;
 public class Project {
 
 	private int projectId;
-	private int managerId;
+	private long managerId;
 	private List<Member> members;
 	private String projectName;
 	private Date startDate;
